Use "Antioch::has_nan()", not direct isnan()
We don't want to conflict with any future isnan() additions to our
non-plain-scalar options.

This also is clearer for vectorized types - isnan(vectorized) should
really return a vectorized-bool, not just a bool.
