+++ src/utilities/include/antioch/eigen_utils.h
@@ -104,7 +104,7 @@ template <typename T>
 inline
 typename Antioch::enable_if_c<is_eigen<T>::value,
   bool>::type
-isnan(const T& in)
+has_nan(const T& in)
 {
   using std::isnan;
