+++ src/utilities/include/antioch/eigen_utils_decl.h
@@ -170,7 +170,7 @@ template <typename T>
 inline
 typename Antioch::enable_if_c<is_eigen<T>::value,
   bool>::type
-isnan(const T& in);
+has_nan(const T& in);
 
 
 template <
