+++ src/utilities/include/antioch/metaphysicl_utils_decl.h
@@ -102,7 +102,7 @@ inline
 typename Antioch::enable_if_c<
   is_metaphysicl<T>::value,
   bool>::type
-isnan (const T& in);
+has_nan (const T& in);
 
 template <typename T>
 struct has_size<T, typename Antioch::enable_if_c<is_metaphysicl<T>::value,void>::type>;
