+++ src/utilities/include/antioch/metaprogramming.h
@@ -35,70 +35,6 @@
 
 namespace Antioch
 {
-#define ANTIOCH_PLAIN_SCALAR(Type) \
- \
-inline \
-Type \
-max (const Type& in) { return in; } \
- \
-inline \
-Type \
-min (const Type& in) { return in; } \
- \
-template <> \
-inline \
-Type if_else(bool condition, \
-             Type if_true, \
-             Type if_false) \
-{ \
-  if (condition) \
-    return if_true; \
-  else \
-    return if_false; \
-} \
- \
-template <> \
-struct has_size<Type> \
-{ \
-  const static bool value = false; \
-}; \
- \
-template <> \
-struct return_auto<Type> \
-{ \
-  const static bool value = false; \
-}; \
- \
-template <> \
-struct value_type<Type> \
-{ \
-  typedef Type type; \
-}; \
- \
-template <> \
-struct raw_value_type<Type> \
-{ \
-  typedef Type type; \
-}; \
- \
-template <typename NewScalar> \
-struct rebind<Type, NewScalar> \
-{ \
-  typedef NewScalar type; \
-}
-
-
-ANTIOCH_PLAIN_SCALAR(float);
-ANTIOCH_PLAIN_SCALAR(double);
-ANTIOCH_PLAIN_SCALAR(long double);
-ANTIOCH_PLAIN_SCALAR(bool);
-ANTIOCH_PLAIN_SCALAR(short);
-ANTIOCH_PLAIN_SCALAR(int);
-ANTIOCH_PLAIN_SCALAR(long);
-ANTIOCH_PLAIN_SCALAR(unsigned short);
-ANTIOCH_PLAIN_SCALAR(unsigned int);
-ANTIOCH_PLAIN_SCALAR(unsigned long);
-
 template <typename T>
 inline
 void set_zero(T& output) { output = 0; }
